build: make `check` cover the workspace, not just the root package
`cargo test` and `cargo clippy` without --workspace run only the root
package, so nothing under crates/ was ever checked here. A failing
eink-dither test still printed "All checks passed!" -- which is how the
dithering work in this branch has looked green throughout. Adding
--all-targets to clippy likewise brings in test files, which were
skipped for the same reason.

Coverage goes from 484 to 951 tests.

Fixes what that exposed: four `&`-on-a-reference calls in git_fetch's
test helpers, and in the dithering diagnostics an `iter().any()` that
is `contains`, a `get(0)`, a redundant i32 cast, and a literal passed
as a format argument.

Also widens the noise_scale sweep from three algorithms to all nine,
which is what showed the optimum is not uniform: the narrow kernels
(Sierra Lite at 3 neighbours, Floyd-Steinberg at 4) have a genuine
optimum and get worse past it, while the wide ones keep improving.

+++ crates/eink-dither/tests/visual_compare.rs
@@ -381,7 +381,7 @@ fn sharp_scene() -> (Vec<Srgb>, usize, usize) {
             for y in 0..4 {
                 for x in 0..3 {
                     if (ch + row + x) % 3 != 0 {
-                        set((bx + x) as i32, (by + y) as i32, black);
+                        set(bx + x, by + y, black);
                     }
                 }
             }
